Move tests from `api_test.go` to `organization_test.go`
